// generated-rest-api-client-typescript-node/model/customUserPreference.ts
/**
 * Papermerge REST API
 * Document management system designed for digital archives
 *
 * The version of the OpenAPI document: 2.1.0b12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CustomUserPreference {
    'section': string;
    'name': string;
    'identifier': string;
    '_default': string;
    'value': string;
    'verboseName': string;
    'helpText': string;
    'additionalData': string;
    'field': string;
    'id': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "section",
            "baseName": "section",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "identifier",
            "baseName": "identifier",
            "type": "string"
        },
        {
            "name": "_default",
            "baseName": "default",
            "type": "string"
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "string"
        },
        {
            "name": "verboseName",
            "baseName": "verbose_name",
            "type": "string"
        },
        {
            "name": "helpText",
            "baseName": "help_text",
            "type": "string"
        },
        {
            "name": "additionalData",
            "baseName": "additional_data",
            "type": "string"
        },
        {
            "name": "field",
            "baseName": "field",
            "type": "string"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CustomUserPreference.attributeTypeMap;
    }
}