// generated-rest-api-client-typescript-node/model/datum.ts
/**
 * Papermerge REST API
 * Document management system designed for digital archives
 *
 * The version of the OpenAPI document: 2.1.0b12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Resource } from './resource';

/**
* singular item
*/
export class Datum {
    'data'?: Resource;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Resource"
        }    ];

    static getAttributeTypeMap() {
        return Datum.attributeTypeMap;
    }
}