// generated-rest-api-client-typescript-node/model/documentDetails.ts
/**
 * Papermerge REST API
 * Document management system designed for digital archives
 *
 * The version of the OpenAPI document: 2.1.0b12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DocumentDetailsData } from './documentDetailsData';

export class DocumentDetails {
    'data'?: DocumentDetailsData;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "DocumentDetailsData"
        }    ];

    static getAttributeTypeMap() {
        return DocumentDetails.attributeTypeMap;
    }
}