// generated-rest-api-client-typescript-node/model/documentDetailsData.ts
/**
 * Papermerge REST API
 * Document management system designed for digital archives
 *
 * The version of the OpenAPI document: 2.1.0b12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DocumentDetailsDataAttributes } from './documentDetailsDataAttributes';
import { DocumentDetailsDataRelationships } from './documentDetailsDataRelationships';

export class DocumentDetailsData {
    'type': DocumentDetailsData.TypeEnum;
    'id'?: string;
    'attributes'?: DocumentDetailsDataAttributes;
    'relationships'?: DocumentDetailsDataRelationships;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "type",
            "baseName": "type",
            "type": "DocumentDetailsData.TypeEnum"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "attributes",
            "baseName": "attributes",
            "type": "DocumentDetailsDataAttributes"
        },
        {
            "name": "relationships",
            "baseName": "relationships",
            "type": "DocumentDetailsDataRelationships"
        }    ];

    static getAttributeTypeMap() {
        return DocumentDetailsData.attributeTypeMap;
    }
}

export namespace DocumentDetailsData {
    export enum TypeEnum {
        Documents = <any> 'documents'
    }
}