// generated-rest-api-client-typescript-node/model/documentDetailsDataAttributes.ts
/**
 * Papermerge REST API
 * Document management system designed for digital archives
 *
 * The version of the OpenAPI document: 2.1.0b12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DocumentVersion } from './documentVersion';

export class DocumentDetailsDataAttributes {
    'id'?: string;
    'title': string;
    'lang'?: string;
    'fileName'?: string;
    'ocr'?: boolean;
    'ocrStatus'?: DocumentDetailsDataAttributes.OcrStatusEnum;
    'versions'?: Array<DocumentVersion>;
    'size'?: number;
    'pageCount'?: number;
    'createdAt'?: Date;
    'updatedAt'?: Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "lang",
            "baseName": "lang",
            "type": "string"
        },
        {
            "name": "fileName",
            "baseName": "file_name",
            "type": "string"
        },
        {
            "name": "ocr",
            "baseName": "ocr",
            "type": "boolean"
        },
        {
            "name": "ocrStatus",
            "baseName": "ocr_status",
            "type": "DocumentDetailsDataAttributes.OcrStatusEnum"
        },
        {
            "name": "versions",
            "baseName": "versions",
            "type": "Array<DocumentVersion>"
        },
        {
            "name": "size",
            "baseName": "size",
            "type": "number"
        },
        {
            "name": "pageCount",
            "baseName": "page_count",
            "type": "number"
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return DocumentDetailsDataAttributes.attributeTypeMap;
    }
}

export namespace DocumentDetailsDataAttributes {
    export enum OcrStatusEnum {
        Unknown = <any> 'unknown',
        Received = <any> 'received',
        Started = <any> 'started',
        Succeeded = <any> 'succeeded',
        Failed = <any> 'failed'
    }
}