// generated-rest-api-client-typescript-node/model/errorSource.ts
/**
 * Papermerge REST API
 * Document management system designed for digital archives
 *
 * The version of the OpenAPI document: 2.1.0b12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ErrorSource {
    /**
    * A [JSON Pointer](https://tools.ietf.org/html/rfc6901) to the associated entity in the request document [e.g. `/data` for a primary data object, or `/data/attributes/title` for a specific attribute.
    */
    'pointer'?: string;
    /**
    * A string indicating which query parameter caused the error.
    */
    'parameter'?: string;
    'meta'?: { [key: string]: any; };

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "pointer",
            "baseName": "pointer",
            "type": "string"
        },
        {
            "name": "parameter",
            "baseName": "parameter",
            "type": "string"
        },
        {
            "name": "meta",
            "baseName": "meta",
            "type": "{ [key: string]: any; }"
        }    ];

    static getAttributeTypeMap() {
        return ErrorSource.attributeTypeMap;
    }
}