// generated-rest-api-client-typescript-node/model/failure.ts
/**
 * Papermerge REST API
 * Document management system designed for digital archives
 *
 * The version of the OpenAPI document: 2.1.0b12
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Jsonapi } from './jsonapi';
import { Link } from './link';
import { ModelError } from './modelError';

export class Failure {
    'errors': Set<ModelError>;
    'meta'?: { [key: string]: any; };
    'jsonapi'?: Jsonapi;
    'links'?: { [key: string]: Link; };

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "errors",
            "baseName": "errors",
            "type": "Set<ModelError>"
        },
        {
            "name": "meta",
            "baseName": "meta",
            "type": "{ [key: string]: any; }"
        },
        {
            "name": "jsonapi",
            "baseName": "jsonapi",
            "type": "Jsonapi"
        },
        {
            "name": "links",
            "baseName": "links",
            "type": "{ [key: string]: Link; }"
        }    ];

    static getAttributeTypeMap() {
        return Failure.attributeTypeMap;
    }
}